lib/js/datachannel_test_2peer/label/008.html

<!DOCTYPE html>
<html lang="en" class="a">
	<head>
		<meta charset="utf-8">
		<!-- Testharness Timeout to long -> 60 Seconds-->
		<meta name="timeout" content="long">
		<title>WebRTC - Label Test</title>
		<script src="../resources/testharness.js"></script>
		<script src="../resources/testharnessreport.js"></script><link href="../resources/testharness.css" type="text/css" rel="stylesheet"/>
		<!-- Including the Firebase.js to send the ICE Candidate to the other peer-->
		<script type="text/javascript" src='https://cdn.firebase.com/v0/firebase.js'></script>
	</head>
	<body class="a">
		<div>
			<div id="signallingInfo"></div>
			
				
			
		</div>
		<hr />

		<!-- Testharness log -->
		<div id="log"></div>
		<script src="../datachannel2peers.js" type="text/javascript"></script>
		<script type="text/javascript">
		// FIXME: FIREFOX Crashes after 8 Channels.
            var label = "test";
            var count = 9;
            var dataChannel = new Array();
            var test = async_test("Set up multiple DataChannels with the same label - try " + count + " DataChannels with same label");
            if (offerer) {
                function createDataChannel(i) {
                    dataChannel[i] = peerConnection.createDataChannel(label);
                    console.log("Create Channel " + i);
                    dataChannel[i].onopen = test.step_func(function() {
                        assert_equals(dataChannel[i].label, label, "Wrong label");
                        i++;
                        if (i < count) {
                            createDataChannel(i);
                        } else {
                            test.done();
                        }
                    });
                }

                createDataChannel(0);
            } else {
                var i = 0;
                peerConnection.ondatachannel = test.step_func(function(e) {
                    dataChannel[i] = e.channel;
                    console.log("Got a Channel " + i);
                    if (i == (count-1)) {
                        assert_equals(dataChannel[i].label, label, "Wrong label");
                        test.done();
                    }
                    i++;
                });
            }

		</script>
	</body>
</html>

